+++ NEWS.md
@@ -9,6 +9,8 @@ https://github.com/surh/HMVAR/issues/6).
 5. Updating `read_midas_data` function, and adding examples.
 6. Function `midas_mktest` now accepts either a file path to map
 or a data frame witht the map. It also has examples.
+7. Added unit testing support via testhat (Issue
+https://github.com/surh/HMVAR/issues/25).
 
 # 0.1-0
 1. Added exectuable script to perform mktest.